+++ CHANGES.rst
@@ -6,6 +6,7 @@ Hiveeyes MicroPython Datalogger CHANGES
 Development
 ===========
 - Upgrade to ``Pycom MicroPython 1.20.0.rc11``
+- Stop leaking sensible information into settings output
 
 
 2019-06-07 0.3.0